+++ test/test_autocli_editmode.sh
@@ -300,7 +300,7 @@ edit table
 show config text
 EOF
 new "show config text"
-expectpart "$(cat $fin | $clixon_cli -f $cfg 2>&1)" 0 "parameter {" "name a;" "value 42;"
+expectpart "$(cat $fin | $clixon_cli -f $cfg 2>&1)" 0 "parameter a {" "value 42;"
 
 cat <<EOF > $fin
 edit table